backend/requirement.txt
fastapi
uvicorn
google-generativeai
pandas
scikit-learn
joblib